Added support for complete match on retract fact
When retracting a fact we should be able to specify either
partial or complete facts. The default is partial since events
can be huge and it might not be practical to specify the whole
fact. But in some cases when the event is small they might want
to do a complete match, but exclude any system added keys like meta.